MediaPlayer NOT WORKING ON MAC?
Okay so basically this is the correct code to make the mediaplayer it's in MyTunesController.
Tried to make it work and play the song together with Nedas, but it doesn't play the sound probably cuz of mac.
So Me and Nedas gave up and I am done someone try it on their amazing Windows and hopefully it works for you ppl. bye§